Refactor: split out logic from LayoutHelper to PreloaderHelper
