+++ tests/Helpers/LayoutHelperTest.php
@@ -4,95 +4,6 @@ use JeroenNoten\LaravelAdminLte\Helpers\LayoutHelper;
 
 class LayoutHelperTest extends TestCase
 {
-    public function testEnableDisablePreloaderFullscreenMode()
-    {
-        // Test with config enabled.
-
-        config(['adminlte.preloader.enabled' => true]);
-
-        $this->assertTrue(LayoutHelper::isPreloaderEnabled());
-        $this->assertTrue(LayoutHelper::isPreloaderEnabled('fullscreen'));
-        $this->assertFalse(LayoutHelper::isPreloaderEnabled('cwrapper'));
-
-        // Test with config disabled.
-
-        config(['adminlte.preloader.enabled' => false]);
-
-        $this->assertFalse(LayoutHelper::isPreloaderEnabled());
-        $this->assertFalse(LayoutHelper::isPreloaderEnabled('fullscreen'));
-        $this->assertFalse(LayoutHelper::isPreloaderEnabled('cwrapper'));
-    }
-
-    public function testEnableDisablePreloaderCWrapperMode()
-    {
-        // Test with config enabled.
-
-        config([
-            'adminlte.preloader.enabled' => true,
-            'adminlte.preloader.mode' => 'cwrapper',
-        ]);
-
-        $this->assertFalse(LayoutHelper::isPreloaderEnabled());
-        $this->assertFalse(LayoutHelper::isPreloaderEnabled('fullscreen'));
-        $this->assertTrue(LayoutHelper::isPreloaderEnabled('cwrapper'));
-
-        // Test with config disabled.
-
-        config([
-            'adminlte.preloader.enabled' => false,
-            'adminlte.preloader.mode' => 'cwrapper',
-        ]);
-
-        $this->assertFalse(LayoutHelper::isPreloaderEnabled());
-        $this->assertFalse(LayoutHelper::isPreloaderEnabled('fullscreen'));
-        $this->assertFalse(LayoutHelper::isPreloaderEnabled('cwrapper'));
-    }
-
-    public function testMakePreloaderClasses()
-    {
-        // Test without config.
-
-        $data = LayoutHelper::makePreloaderClasses();
-        $this->assertEquals(
-            'preloader flex-column justify-content-center align-items-center',
-            $data
-        );
-
-        // Test with cwrapper mode enabled.
-
-        config([
-            'adminlte.preloader.enabled' => true,
-            'adminlte.preloader.mode' => 'cwrapper',
-        ]);
-
-        $data = LayoutHelper::makePreloaderClasses();
-
-        $this->assertStringContainsString(
-            'preloader flex-column justify-content-center align-items-center',
-            $data
-        );
-
-        $this->assertStringContainsString('position-absolute', $data);
-    }
-
-    public function testMakePreloaderStyle()
-    {
-        // Test without config.
-
-        $data = LayoutHelper::makePreloaderStyle();
-        $this->assertEquals('', $data);
-
-        // Test with cwrapper mode enabled.
-
-        config([
-            'adminlte.preloader.enabled' => true,
-            'adminlte.preloader.mode' => 'cwrapper',
-        ]);
-
-        $data = LayoutHelper::makePreloaderStyle();
-        $this->assertStringContainsString('z-index:', $data);
-    }
-
     public function testMakeContentWrapperClasses()
     {
         // Test without config.
